[Backend][SystemC] Rename and move PortDirection and Hierarchy to ir_utils.h
